Eliminate symbol duplication
